feat(): theme-related - templatize .less files to reference main theme object
DRAFT!

First. Clearly we should move away from `less` and commit to emotion/antd for theming

Now deleting the less files is going to be difficult. In the meantime, I wanted to provide a way for less files to source from the main theme object. I decided to go with handlebars since that should be part of the build process.

Considerations:
- this is temporary, goal is to get rid of less altogether. Having this enables layering the theming work, and allows to make the theme DRY, instead of referncing the same colors a bunch of times
- eventually could remove the `.less` files from the repo, and make sure they are dynamically generated on every builds. In the meantime I thought I'd leave them here, and we can instruct people to alter `.less.hds` files and run `npm run compile-less` on demand, if/when altering the main theme